Six month calculation fix and inject timestamp into {{Db-g13}}
Now correctly calculates if last edit was more than six months ago regardless of number of days or milliseconds...
Tags with `{{Db-g13|ts=`timestamp`}}` now.  This will allow a new box to be introduced to the `{{Db-g13}}` template that lets the admin know quickly if the draft is indeed eligible, who nominated it, and if it was a premature nomination.