Significant speed improvements in cleanup() and related code
afcHelper_cleanup() now accepts an additional `type` parameter. If this
parameter is set to "initial," cleanup will *not* run external cleanup
scripts (we set it to "initial" when we're initially launching the
script -- because cleanup is run again before saving, there's no point
in running the external scripts twice). This is in and of itself quite
a large speed improvment. If the parameter is set to "accept," a
variety of accept-only cleanup will occur, and cleanup that was (in old
versions of the code) subsequently "undone" by accept() will simply
never execute.