Fix flawed logic
@Technical-13, remember, the `if` statements execute in order (and run
the *first* one that evaluates to true) -- so in your code, if
`(sub_m.holdrat == 'afd') && (sub_m.addloc)` were both true then it
would execute *that* statement, completely ignoring whether or not
`sub_m.addtl` was true, since it is only checked for in the `if`
statement below that (which never executes). Regardless, fixed. See
test edit
https://en.wikipedia.org/w/index.php?title=Wikipedia:Files_for_upload&di
ff=574868342&oldid=574863486