Remove apparently broken AutoEd implementation as well
importScriptURI is asynchronous, which means the code doesn't wait for
the scripts to load before continuing, resulting in reference errors
(the reason I didn't see them is becuause I had AutoEd installed
already as a userscript, so it was pre-loading it). @wikipedia-mabdul
we need to get away from importScriptURI… and this needs to be figured
out by next release.