Feature: check validity of move targets as they are typed
When accepting a submission and typing a target title, users will be
told if the title is available, if a page already exists at that title,
or, even more specifically, if the title is a redirect (and then in
turn what its target is). Uses a new `afcHelper_checkTarget()` function
and `onkeyup`.